pyproject.toml
[build-system]
requires = ["setuptools","setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "pyspextool"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
    "astropy",
    "scipy",
    "matplotlib",
    "numpy <2.0",
    "pandas",
    "importlib_resources; python_version < '3.11'",
    "git-lfs",
    "astroquery >0.4.8",
    "specutils",
    "dust_extinction",
    "jinja2",
    "pooch",
    "pyyaml",
]
dynamic = ["version"]

[project.optional-dependencies]
test = [
    "pytest",
    "ruff",
    "pytest-cov",
    ]

[tool.setuptools]
include-package-data = true


[tool.setuptools.package-data]
"*" = ["*.dat","*.fits","*.csv","*.txt","*.css","*.yaml"] # includes the data files in instrument_data/ and data/ and batch/

[tool.setuptools_scm]

[tool.pytest.ini_options]
markers = [
    "slow: mark test as slow",
]

[tool.coverage.run]
source = ["pyspextool"]
branch = true

[tool.coverage.report]
show_missing = true
skip_empty = true